VideoPlayer dependency isolation
